`compaction`: port `storage::compaction_config` and `copy_data_segment_reducer::stats`
