+++ src/v/compaction/BUILD
@@ -45,3 +45,27 @@ redpanda_cc_library(
         "@seastar",
     ],
 )
+
+redpanda_cc_library(
+    name = "types",
+    srcs = [
+        "types.cc",
+    ],
+    hdrs = [
+        "types.h",
+    ],
+    implementation_deps = [
+        "//src/v/utils:to_string",
+        "@fmt",
+    ],
+    include_prefix = "compaction",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":key_offset_map",
+        "//src/v/base",
+        "//src/v/model",
+        "//src/v/storage:file_sanitizer_types",
+        "//src/v/storage:scoped_file_tracker",
+        "@seastar",
+    ],
+)
