+++ src/v/storage/disk_log_impl.cc
@@ -496,7 +496,7 @@ disk_log_impl::request_eviction_until_offset(model::offset max_offset) {
 }
 
 ss::future<compaction_result> disk_log_impl::segment_self_compact(
-  compaction_config cfg,
+  compaction::compaction_config cfg,
   ss::lw_shared_ptr<segment> seg,
   bool force_compaction) {
     co_return co_await storage::internal::self_compact_segment(
@@ -513,7 +513,7 @@ ss::future<compaction_result> disk_log_impl::segment_self_compact(
 
 ss::future<> disk_log_impl::adjacent_merge_compact(
   segment_set& segments,
-  compaction_config cfg,
+  compaction::compaction_config cfg,
   std::optional<model::offset> new_start_offset) {
     vlog(
       gclog.trace,
@@ -596,7 +596,8 @@ ss::future<> disk_log_impl::adjacent_merge_compact(
 }
 
 segment_set disk_log_impl::find_sliding_range(
-  const compaction_config& cfg, std::optional<model::offset> new_start_offset) {
+  const compaction::compaction_config& cfg,
+  std::optional<model::offset> new_start_offset) {
     if (
       _last_compaction_window_start_offset.has_value()
       && (_last_compaction_window_start_offset.value()
@@ -648,7 +649,8 @@ segment_set disk_log_impl::find_sliding_range(
 }
 
 ss::future<bool> disk_log_impl::sliding_window_compact(
-  const compaction_config& cfg, std::optional<model::offset> new_start_offset) {
+  const compaction::compaction_config& cfg,
+  std::optional<model::offset> new_start_offset) {
     vlog(gclog.debug, "[{}] running sliding window compaction", config().ntp());
     auto segs = find_sliding_range(cfg, new_start_offset);
     if (segs.empty()) {
@@ -863,7 +865,8 @@ ss::future<bool> disk_log_impl::sliding_window_compact(
 std::optional<
   chunked_vector<std::pair<segment_set::iterator, segment_set::iterator>>>
 disk_log_impl::find_adjacent_compaction_ranges(
-  const compaction_config& cfg, std::optional<model::offset> new_start_offset) {
+  const compaction::compaction_config& cfg,
+  std::optional<model::offset> new_start_offset) {
     {
         // Early return if cluster configured values effectively disables
         // adjacent merge compaction.
@@ -1002,7 +1005,7 @@ disk_log_impl::find_adjacent_compaction_ranges(
 
 ss::future<std::optional<chunked_vector<compaction_result>>>
 disk_log_impl::compact_adjacent_segment_ranges(
-  storage::compaction_config cfg,
+  compaction::compaction_config cfg,
   std::optional<model::offset> new_start_offset) {
     chunked_vector<compaction_result> rs;
     if (auto ranges = find_adjacent_compaction_ranges(cfg, new_start_offset);
@@ -1046,7 +1049,7 @@ disk_log_impl::compact_adjacent_segment_ranges(
 
 ss::future<compaction_result> disk_log_impl::do_compact_adjacent_segments(
   chunked_vector<ss::lw_shared_ptr<segment>>& segments,
-  storage::compaction_config cfg) {
+  compaction::compaction_config cfg) {
     // This shouldn't be the case for any ranges returned from
     // find_adjacent_compaction_ranges(), but it is checked regardless.
     if (segments.size() < 2) {
@@ -1367,7 +1370,7 @@ ss::future<> disk_log_impl::housekeeping(housekeeping_config cfg) {
 }
 
 ss::future<> disk_log_impl::do_compact(
-  compaction_config compact_cfg,
+  compaction::compaction_config compact_cfg,
   std::optional<model::offset> new_start_offset) {
     compact_cfg.asrc = &_compaction_as;
 
@@ -1413,7 +1416,7 @@ ss::future<> disk_log_impl::do_compact(
 }
 
 ss::future<bool> disk_log_impl::chunked_sliding_window_compact(
-  const compaction_config& compact_cfg,
+  const compaction::compaction_config& compact_cfg,
   const segment_set& segs,
   compaction::key_offset_map& map) {
     // The last segment in the segment set is the first segment we attempted and
@@ -1544,7 +1547,7 @@ ss::future<bool> disk_log_impl::chunked_sliding_window_compact(
 }
 
 ss::future<> disk_log_impl::rewrite_segment_with_offset_map(
-  const compaction_config& cfg,
+  const compaction::compaction_config& cfg,
   ss::lw_shared_ptr<segment> seg,
   compaction::key_offset_map& map,
   bool is_finished_window_compaction,
