+++ src/v/storage/tests/compaction_fuzz_test.cc
@@ -159,7 +159,8 @@ ss::future<ot_state> arrange_and_compact(
                 co_await b1.get_disk_log_impl().force_roll();
             }
         }
-        auto compact_cfg = storage::compaction_config(
+        ss::abort_source as;
+        auto compact_cfg = compaction::compaction_config(
           batches.back().last_offset(), std::nullopt, std::nullopt, as);
         std::ignore = co_await b1.apply_sliding_window_compaction(compact_cfg);
         co_await b1.apply_adjacent_merge_compaction(compact_cfg);
