+++ src/v/storage/tests/log_truncate_test.cc
@@ -557,7 +557,7 @@ TEST_F(storage_test_fixture, test_concurrent_truncate_and_compaction) {
     // compaction, which initially self compacts one segment at a time, while
     // leaving room for further windowed compaction.
     ss::abort_source as;
-    compaction_config compaction_cfg(
+    compaction::compaction_config compaction_cfg(
       model::offset::max(), std::nullopt, std::nullopt, as);
     auto& disk_log = *dynamic_cast<disk_log_impl*>(log.get());
     disk_log.adjacent_merge_compact(disk_log.segments(), compaction_cfg).get();
